+++ CHANGELOG.md
@@ -23,8 +23,10 @@ All notable changes to the "cosmy-wasmy" extension will be documented in this fi
 
 ### Added
 
-- Add export saved state to json. Exports the accounts, contracts and ahy historical queries stored.
+- Add export saved state to json. Exports the accounts, contracts and any historical queries stored.
 - Shows a plug icon next to contracts which were imported for the currently active chain config.
+- Contract context menu actions to update contract admin and to remove contract admin
+- Added [#10](https://github.com/spoo-bar/cosmy-wasmy/issues/10) - Developers can add small notes and comments to their imported contracts. The added notes are shown on hovering over the contracts in the sidebar contract view. The notes supports markdown. Note : The added notes are only for the dev view, this info is  not stored on chain.
 
 ### Changed